@utkarsh530 2024-02-26
Before:

```
julia> @benchmark sol = solve(prob,
           ParallelSyncPSOKernel(1000, backend = CUDA.CUDABackend()),
           maxiters = 500)
BenchmarkTools.Trial: 87 samples with 1 evaluation.
 Range (min … max):  49.353 ms … 187.970 ms  ┊ GC (min … max): 0.00% … 38.72%
 Time  (median):     53.320 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   57.881 ms ±  23.307 ms  ┊ GC (mean ± σ):  3.74% ±  6.52%

  ▇█▃                                                           
  ███▇▄▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▃ ▁
  49.4 ms         Histogram: frequency by time          179 ms <

 Memory estimate: 3.00 MiB, allocs estimate: 71083.
```

After:

```
julia> @benchmark solve(prob,
           ParallelSyncPSOKernel(1000, backend = CUDA.CUDABackend()),
           maxiters = 500)
BenchmarkTools.Trial: 132 samples with 1 evaluation.
 Range (min … max):  34.288 ms … 166.877 ms  ┊ GC (min … max): 0.00% … 22.80%
 Time  (median):     34.847 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   38.072 ms ±  19.359 ms  ┊ GC (mean ± σ):  2.52% ±  3.89%

  █                                                             
  █▆▄▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▄ ▄
  34.3 ms       Histogram: log(frequency) by time       166 ms <

 Memory estimate: 1.71 MiB, allocs estimate: 41525.
```